moving one segmentation method to new classes. cleaning up code elsewhere such as incomplete parameterization declarations for use. removing deprecated unused junit imports.

+++ src/algorithms/imageProcessing/ImageIOHelper.java
@@ -908,7 +908,7 @@ int z1 = 1;
                 if ((yy < 0) || (yy > (h - 1))) {
                     continue;
                 }
-                input.setRGB((int)xx, (int)yy, rClr, gClr, bClr);
+                input.setRGB(xx, yy, rClr, gClr, bClr);
             }
         }
     }
